requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.16.0
construct==2.10.70
lxml==5.1.0
psutil==5.9.8
pycparser==2.21
pycryptodomex==3.20.0
pykeepass==4.0.6
python-dateutil==2.8.2
python-gnupg==0.5.2
six==1.16.0
